Make is simpler
