+++ etc/dolphin/dolphin.conf
@@ -1,6 +1,6 @@
 [DEFAULT]
 api_paste_config = /etc/dolphin/api-paste.ini
-dolphin_cryptor = base64
+dolphin_cryptor = dolphin.cryptor._Base64
 
 [database]
 connection = sqlite:////var/lib/dolphin/dolphin.sqlite
